+++ wallet-gateway/extension/entrypoints/content/index.ts
@@ -1,0 +1,14 @@
+// Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import { announceSelf } from './announce-self'
+import { jsonRpcProxy } from './json-rpc-proxy'
+
+export default defineContentScript({
+    matches: ['file://*/*', 'http://*/*', 'https://*/*'],
+    main() {
+        // initialize content script
+        announceSelf()
+        jsonRpcProxy()
+    },
+})
